Move NavBar component to navBar folder
